Update 0.1.22
Fixed:
* Reading from image and calculating brightness difference
* Outputting image in ascii
Improved:
* Code readability
* Optimalization

+++ src/consoleUtils/UnixConsole.h
@@ -2,7 +2,7 @@
 // Name        : UnixConsole
 // Author      : Riyufuchi
 // Created on  : 20.07.2020
-// Last Edit   : 22.11.2022
+// Last Edit   : 06.03.2023
 // Description : This class contains methods for working with a Linux console
 //============================================================================
 
@@ -14,7 +14,7 @@
 class UnixConsole
 {
 public:
-	struct Color
+	struct Color // This way, there is no need to cast, because uint8_t would be treated as char not numerical value
 	{
 		short int red;
 		short int blue;
@@ -24,7 +24,7 @@ public:
 	void writeText(const char* text);
 	void writeText(std::string text);
 	void writeText(Color color, const char* text);
-	void writeText(int  r, int g, int b, const char* text);
+	void writeText(short int r, short int g, short int b, const char* text);
 	void setTextColor(Color color);
 	Color newColor(short int r, short int g, short int b);
 	~UnixConsole();
